+++ app/Http/Controllers/FirstdataDebitoController.php
@@ -1,0 +1,85 @@
+<?php
+
+namespace App\Http\Controllers;
+
+use App\Models\Firstdata_debito;
+use Illuminate\Http\Request;
+
+class FirstdataDebitoController extends Controller
+{
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function index()
+    {
+        //
+    }
+
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function create()
+    {
+        //
+    }
+
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\Response
+     */
+    public function store(Request $request)
+    {
+        //
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  \App\Models\Firstdata_debito  $firstdata_debito
+     * @return \Illuminate\Http\Response
+     */
+    public function show(Firstdata_debito $firstdata_debito)
+    {
+        //
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  \App\Models\Firstdata_debito  $firstdata_debito
+     * @return \Illuminate\Http\Response
+     */
+    public function edit(Firstdata_debito $firstdata_debito)
+    {
+        //
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  \App\Models\Firstdata_debito  $firstdata_debito
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, Firstdata_debito $firstdata_debito)
+    {
+        //
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\Models\Firstdata_debito  $firstdata_debito
+     * @return \Illuminate\Http\Response
+     */
+    public function destroy(Firstdata_debito $firstdata_debito)
+    {
+        //
+    }
+}