+++ database/migrations/2023_05_10_130305_create_conceptos_debitos_table.php
@@ -1,0 +1,33 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class CreateConceptosDebitosTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::create('conceptos_debitos', function (Blueprint $table) {
+            $table->id();
+            $table->timestamps();
+            $table->string('concepto', 30);
+            $table->boolean('desactivado');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::dropIfExists('conceptos_debitos');
+    }
+}