+++ CHANGELOG.md
@@ -30,8 +30,11 @@ ___ -->
 
 ## [0.8.10] - 2023-05-10
 ### Add
+#### Composer Updste REQUIRED
 1.  Vista de actualizar Calles cargando un archivo.
-2.  ABM de BTF Debitos.
+2.  ABM de Conceptos Debitos.
+3.  ABM de BTF Debitos.
+4.  Se instala DomPDF
 
 ## [0.8.9] - 2023-05-06
 ### Changed